LIR serializer refactor
